ExceptionMapper e documentação

+++ src/main/java/quarkus/crud/rest/FuncionalidadeRest.java
@@ -3,11 +3,30 @@ package quarkus.crud.rest;
 import io.quarkus.panache.common.Sort;
 import jakarta.transaction.Transactional;
 import jakarta.ws.rs.*;
+import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
+import org.jboss.resteasy.reactive.common.util.RestMediaType;
 import quarkus.crud.entity.Funcionalidade;
 import quarkus.crud.repository.FuncionalidadeDAO;
 
 import java.util.List;
+
+
+/**
+ *
+ *      Media Type (@Consumes e @Produces):
+ *         - When a JSON extension is installed such as quarkus-resteasy-jackson or quarkus-resteasy-jsonb
+ *           * Object:  application/json
+ *           * String:  text/plain
+ *           * File: application/octet-stream
+ *
+ *
+ *      HAL+JSON
+ *          * https://quarkus.io/guides/resteasy-reactive#web-links-support
+ *
+ *
+ */
+
 
 @Path("/funcionalidade")
 public class FuncionalidadeRest {
@@ -22,10 +41,20 @@ public class FuncionalidadeRest {
     }
 
     @GET
+    @Produces({MediaType.APPLICATION_JSON, RestMediaType.APPLICATION_HAL_JSON})
     public List<Funcionalidade> listAll() {
 
         return dao.listAll(Sort.by("id", Sort.Direction.Descending, Sort.NullPrecedence.NULLS_LAST));
     }
+
+
+    @GET
+    @Path("/count")
+    public Long count() {
+
+        return dao.count();
+    }
+
 
     @GET
     @Path("{id}")
@@ -40,7 +69,7 @@ public class FuncionalidadeRest {
     public Funcionalidade add(Funcionalidade f) {
 
         if(f.id != null)
-            throw new WebApplicationException("Somente entidades com id nulo podem ser adicionadas");
+            throw new WebApplicationException("Somente entidades com id nulo podem ser adicionadas", 500);
 
         dao.persist(f);
 
@@ -49,12 +78,13 @@ public class FuncionalidadeRest {
 
 
     @PUT
-    @Path("{id}")
+    @Path("{id}")     //@Consumes(MediaType.APPLICATION_JSON) @Produces(MediaType.APPLICATION_JSON)
     @Transactional
     public Funcionalidade update(Long id, Funcionalidade toSave) {
 
-        Funcionalidade f = dao.findByIdOptional(id)
-                .orElseThrow(() -> new WebApplicationException("Entidade com id " + id + " não existe.", 404));
+
+        if(!dao.existe(id))
+            throw new WebApplicationException("Entidade com id " + id + " não existe.", 404);
 
         return dao.getEntityManager().merge(toSave);
     }
@@ -65,8 +95,8 @@ public class FuncionalidadeRest {
     @Transactional
     public Response delete(Long id) {
 
-        Funcionalidade f = dao.findByIdOptional(id)
-                .orElseThrow(() -> new WebApplicationException("Entidade com id " + id + " não existe.", 404));
+        if(!dao.existe(id))
+            throw new WebApplicationException("Entidade com id " + id + " não existe.", 404);
 
         dao.deleteById(id);
